+++ src/lib.rs
@@ -129,6 +129,9 @@ mod shape;
 use self::shape_plan_cache::*;
 mod shape_plan_cache;
 
+pub use self::shape_run_cache::*;
+mod shape_run_cache;
+
 #[cfg(feature = "swash")]
 pub use self::swash::*;
 #[cfg(feature = "swash")]